requirements.txt
djangorestframework
drf-yasg
django-filter
django-environ

psycopg2-binary==2.8.6

debugpy
django-cors-headers
